Data: move Queries classes to Queries directory
