+++ inc/Data/Queries/FloodQueries.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Vichan\Data;
+namespace Vichan\Data\Queries;
 
 /**
  * Handles direct database interactions for the `flood` table.
